+++ src/byteutils.h
@@ -47,25 +47,25 @@ void push_short(std::vector<uint8_t>& stack, const uint16_t& value);
 void push_int(std::vector<uint8_t>& stack, const uint32_t& value);
 void push_long(std::vector<uint8_t>& stack, const uint64_t& value);
 
-uint8_t read_byte(const uint8_t* stack, uint32_t index);
-uint16_t read_short(const uint8_t* stack, uint32_t index);
-uint32_t read_int(const uint8_t* stack, uint32_t index);
-uint64_t read_long(const uint8_t* stack, uint32_t index);
+uint8_t read_byte(const uint8_t* stack, uint64_t index);
+uint16_t read_short(const uint8_t* stack, uint64_t index);
+uint32_t read_int(const uint8_t* stack, uint64_t index);
+uint64_t read_long(const uint8_t* stack, uint64_t index);
 
-uint8_t read_byte(const std::vector<uint8_t>& stack, uint32_t index);
-uint16_t read_short(const std::vector<uint8_t>& stack, uint32_t index);
-uint32_t read_int(const std::vector<uint8_t>& stack, uint32_t index);
-uint64_t read_long(const std::vector<uint8_t>& stack, uint32_t index);
+uint8_t read_byte(const std::vector<uint8_t>& stack, uint64_t index);
+uint16_t read_short(const std::vector<uint8_t>& stack, uint64_t index);
+uint32_t read_int(const std::vector<uint8_t>& stack, uint64_t index);
+uint64_t read_long(const std::vector<uint8_t>& stack, uint64_t index);
 
-void write_byte(uint8_t* stack, uint32_t index, uint8_t value);
-void write_short(uint8_t* stack, uint32_t index, uint16_t value);
-void write_int(uint8_t* stack, uint32_t index, uint32_t value);
-void write_long(uint8_t* stack, uint32_t index, uint64_t value);
+void write_byte(uint8_t* stack, uint64_t index, uint8_t value);
+void write_short(uint8_t* stack, uint64_t index, uint16_t value);
+void write_int(uint8_t* stack, uint64_t index, uint32_t value);
+void write_long(uint8_t* stack, uint64_t index, uint64_t value);
 
-void write_byte(std::vector<uint8_t>& stack, uint32_t index, uint8_t value);
-void write_short(std::vector<uint8_t>& stack, uint32_t index, uint16_t value);
-void write_int(std::vector<uint8_t>& stack, uint32_t index, uint32_t value);
-void write_long(std::vector<uint8_t>& stack, uint32_t index, uint64_t value);
+void write_byte(std::vector<uint8_t>& stack, uint64_t index, uint8_t value);
+void write_short(std::vector<uint8_t>& stack, uint64_t index, uint16_t value);
+void write_int(std::vector<uint8_t>& stack, uint64_t index, uint32_t value);
+void write_long(std::vector<uint8_t>& stack, uint64_t index, uint64_t value);
 
 }
